backend/requirements.txt
Django==3.2.13
environs==7.3.1
GitPython
gunicorn
djangorestframework
numpy
psycopg2
django-cors-headers
django-celery-beat
celery
redis
requests
ghlinguist
pytest
pytest-django
pytest-cov
coverage